fix(perks): standardize data types for settings and improve perk state management
- Updated integer-to-long conversions for perk settings to ensure consistent data types (e.g., trail interval, duration, charge ticks).
- Fixed `ProximityMeterPerk` to properly handle null values in `proximityMeterNearest` map.
- Adjusted particle spawning logic in `BlockSwapItem` for enhanced visual effects.

+++ plugin/src/main/java/de/thecoolcraft11/hideAndSeek/perk/impl/hider/ShadowStepPerk.java
@@ -80,7 +80,7 @@ public class ShadowStepPerk extends BasePerk {
 
             plugin.getPerkStateManager().cancelTask(player, getId());
 
-            int chargeTicks = plugin.getSettingRegistry().get("perks.perk.hider_shadow_step.charge-ticks", 30);
+            long chargeTicks = plugin.getSettingRegistry().get("perks.perk.hider_shadow_step.charge-ticks", 30L);
             BukkitTask chargeTask = Bukkit.getScheduler().runTaskLater(plugin, () -> {
                 plugin.getPerkStateManager().shadowStepChargeTask.remove(player.getUniqueId());
                 if (!player.isOnline() || player.getGameMode() == GameMode.SPECTATOR || player.isDead()) {
